refactors functions into functions.php

+++ mysql/functions.php
@@ -4,9 +4,9 @@ include 'db.php';
 
 function createRows() {
   if(isset($_POST['submit'])) {
-  global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
 
     $query = "INSERT INTO users(username, password) ";
     $query .= "VALUES ('$username', '$password')";
@@ -40,34 +40,42 @@ function showAllData() {
 
 
 function updateTable() {
-  global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
-  $id = $_POST['id'];
+  if(isset($_POST['submit'])) {
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+    $id = $_POST['id'];
 
-  $query = "UPDATE users SET ";
-  $query .= "username = '$username', ";
-  $query .= "password = '$password' ";
-  $query .= "WHERE id = $id";
+    $query = "UPDATE users SET ";
+    $query .= "username = '$username', ";
+    $query .= "password = '$password' ";
+    $query .= "WHERE id = $id";
 
-  $result = mysqli_query($connection, $query);
-  if(!$result) {
-    die("Query Failed" . mysqli_error($connection));
+    $result = mysqli_query($connection, $query);
+    if(!$result) {
+      die("Query Failed" . mysqli_error($connection));
+    } else {
+      echo "User Updated!";
+    }
   }
 }
 
 function deleteRows() {
-  global $connection;
-  $username = $_POST['username'];
-  $password = $_POST['password'];
-  $id = $_POST['id'];
+  if(isset($_POST['submit'])) {
+    global $connection;
+    $username = $_POST['username'];
+    $password = $_POST['password'];
+    $id = $_POST['id'];
 
-  $query = "DELETE FROM users ";
-  $query .= "WHERE id = $id";
+    $query = "DELETE FROM users ";
+    $query .= "WHERE id = $id";
 
-  $result = mysqli_query($connection, $query);
-  if(!$result) {
-    die("Query Failed" . mysqli_error($connection));
+    $result = mysqli_query($connection, $query);
+    if(!$result) {
+      die("Query Failed" . mysqli_error($connection));
+    } else {
+      echo "User Deleted!";
+    }
   }
 }
